+++ modules/core/osgi-service-tracker-collections/src/main/java/com/liferay/osgi/service/tracker/collections/internal/osgi/util/tracker/DefaultServiceTrackerCustomizer.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.osgi.service.tracker.collections.internal;
+package com.liferay.osgi.service.tracker.collections.internal.osgi.util.tracker;
 
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.ServiceReference;
